Worldgen: fix overworld depth (offset constant + ridges_folded constants)
Two 1:1 fixes that make the router `depth` function bit-exact vs Java:
- offset uses Java's GLOBAL_OFFSET constant(-0.50375F): a float literal widened
  to double (-0.5037500262260437), not the exact double -0.50375.
- peaksAndValleys (RIDGES_FOLDED) used the float helper's 0.6666667/0.33333334;
  the density-function version is double arithmetic with -2/3 and -1/3
  (mul(add(abs(add(abs(w),-2/3)),-1/3),-3)).

density_parity: 56 -> 16 mismatches; temperature/vegetation/continents/erosion/
ridges/depth now all bit-exact. Remaining 16 are final_density only (a separate
sub-function divergence, not low-bit).

+++ mcpp/src/world/level/levelgen/NoiseRouterData.cpp
@@ -294,7 +294,9 @@ NoiseRouter overworld(RandomState& randomState, bool largeBiomes, bool amplified
     auto ridgesCoordinate = densityCoordinate(ridgesFolded);
 
     auto offset = DensityFunctions::add(
-        DensityFunctions::constant(-0.50375),
+        // Java: DensityFunctions.constant(-0.50375F) — a float literal widened to
+        // double (= -0.5037500262260437), NOT the exact double -0.50375.
+        DensityFunctions::constant(static_cast<double>(-0.50375f)),
         spline(TerrainProvider::overworldOffset(continentsCoordinate, erosionCoordinate, ridgesCoordinate, amplified)));
     auto factor = spline(TerrainProvider::overworldFactor(continentsCoordinate, erosionCoordinate, weirdnessCoordinate, ridgesCoordinate, amplified));
     auto depth = offsetToDepth(offset);
